consumer-agent/requirements.txt
a2a-sdk[http-server]
langgraph>=0.3.18
langchain-anthropic>=0.3.0
httpx>=0.27.0
